Fix remaining clippy warnings (CI green)
- tests.rs: move #![cfg(test)] before doc comment (inner attr order)
- fec/codec.rs: prefix unused `len` with _
- transport/connection.rs: on_packet takes &mut [u8] not &mut Vec<u8>
- benches/fec.rs: needless_range_loop → iter().enumerate()

+++ src/transport/chaff.rs
@@ -94,7 +94,7 @@ impl ChaffScheduler {
         }
         self.lcg = self.lcg.wrapping_mul(6364136223846793005).wrapping_add(1);
         // Jitter ≤ 5ms — small enough not to hurt latency, large enough to blur timing
-        let jitter_us = (self.lcg >> 32) as u64 % 5_000;
+        let jitter_us = (self.lcg >> 32) % 5_000;
         Duration::from_micros(jitter_us)
     }
 }
